Initial commit & testing third party libraries.
Set up the project and testing to make sure third party technology is
working as intended. Usually, I wouldn't include the filters into
version control, as I think it is better for developers to organize
thier solution in the way they prefer; however, it is useful in this
instance to demonstrate my process.