Adding CoinDispenser and tests. Refactor on currency and the CurrencyTakenEvent
The refactor on currency is not nessasary for the project with just
coins, but it would be a serious scalablity problem going forward to add
paper money; therefore, I did the refactor now. Tests remain mostly
unchanged, as functionality wise, the changes were only additive. Make
change story is now fully implemented.

Question for customer going forward: How is change handled with paper
bills and exact change only and returning the customer's money?