Standardized manager ctors. Added asserts.
Moved all the constructors that took an istream to take a Json::Value
reference to move responsiblity to the users to validate the stream.
(Also better for future tests)

Removed throws and added asserts to verify incoming data as it isn't a
recoverable exception so it made no sense to keep it, it should be
assumed the data is valid w/o checking when the application goes live.